+++ include/SplashScreen.hpp
@@ -1,0 +1,8 @@
+#ifndef SPLASHSCREEN_HPP
+#define SPLASHSCREEN_HPP
+
+#include "main.hpp"
+
+void splashScreen();
+
+#endif // SPLASHSCREEN_HPP
